Forgot to teach conservative GC about jsid type
UltraBlame original commit: a69154441de4e1aed24e1fd3d76f38c614f5f9ce

+++ js/src/jsxml.h
@@ -171,9 +171,8 @@ struct JSXML {
     void                *domnode;       
     JSXML               *parent;
     JSObject            *name;
-    
-    size_t              xml_class;      
-    size_t              xml_flags;      
+    uint32              xml_class;      
+    uint32              xml_flags;      
     union {
         JSXMLListVar    list;
         JSXMLElemVar    elem;
